Refactoring explorer view to listeners
