+++ src/listeners/dashboard/BaseListener.ts
@@ -1,7 +1,7 @@
-import { Dashboard } from "../commands/Dashboard";
-import { DashboardCommand } from "../dashboardWebView/DashboardCommand";
-import { DashboardMessage } from "../dashboardWebView/DashboardMessage";
-import { Logger } from "../helpers/Logger";
+import { Dashboard } from "../../commands/Dashboard";
+import { DashboardCommand } from "../../dashboardWebView/DashboardCommand";
+import { DashboardMessage } from "../../dashboardWebView/DashboardMessage";
+import { Logger } from "../../helpers/Logger";
 
 
 export abstract class BaseListener {
