+++ src/models/index.ts
@@ -2,6 +2,9 @@ export * from './Choice';
 export * from './ContentFolder';
 export * from './CustomTaxonomyData';
 export * from './DashboardData';
+export * from './DataFile';
+export * from './DataFolder';
+export * from './DataType';
 export * from './DraftField';
 export * from './Framework';
 export * from './MediaPaths';
